Fixed broken command

+++ cogs/filter.py
@@ -6,38 +6,38 @@ class Filter(commands.Cog):
     def __init__(self, bot) -> None:
         self.bot = bot
         self.badwords = [
-            " fuck ",
-            " motherfucker ",
-            " mugi ",
-            " muji ",
-            " fucking ",
-            " motherfucking ",
-            " dick ",
-            " machikne ",
-            " shit ",
-            " tits ",
-            " bitch ",
-            " whore ",
-            " hoe ",
-            " nigga ",
-            " nigger ",
-            " cunt ",
-            " faggot ",
-            " pussy ",
-            " asshole ",
-            " dickhead ",
-            " bastard ",
-            " wanker ",
-            " randi ",
-            " radi ",
-            " chikne ",
-            " chickney ",
-            " gede ",
-            " lado ",
-            " suck my ",
-            " lick my ",
-            " bhosdi ",
-            " madarchut ",
+            "fuck",
+            "motherfucker",
+            "mugi",
+            "muji",
+            "fucking",
+            "motherfucking",
+            "dick",
+            "machikne",
+            "shit",
+            "tit",
+            "bitch",
+            "whore",
+            "hoe",
+            "nigga",
+            "nigger",
+            "cunt",
+            "faggot",
+            "pussy",
+            "asshole",
+            "dickhead",
+            "bastard",
+            "wanker",
+            "randi",
+            "radi",
+            "chikne",
+            "chickney",
+            "gede",
+            "lado",
+            "suck my",
+            "lick my",
+            "bhosdi",
+            "madarchut",
         ]
 
     @commands.Cog.listener()
@@ -45,13 +45,16 @@ class Filter(commands.Cog):
         for word in self.badwords:
             # if word.islower() or word.isupper():
             if message.content.lower().find(word.lower()) != -1:
-                await message.delete()
-                em = discord.Embed(
-                    title="**—————————————WARNING—————————————**",
-                    description=f"{message.author.name}!! You cannot use that word here",
-                    color=discord.Color.purple(),
-                )
-                await message.channel.send(embed=em)
+                if "https://" in message.content:
+                    return
+                else:
+                    await message.delete()
+                    em = discord.Embed(
+                        title="**—————————————WARNING—————————————**",
+                        description=f"{message.author.name}!! You cannot use that word here",
+                        color=discord.Color.purple(),
+                    )
+                    await message.channel.send(embed=em)
 
 
 def setup(bot):
